requirements.txt
boto3==1.20.3
discord.py==1.7.3
